requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv